test: update test
